prometheus_mp: give the on-disk store format its own header
mp_store.cpp carried four things: the byte layout, the file naming, the
writer and the reader. The layout is the one part two differently-built
processes must agree on and the one part that cannot change without a
schema bump, so it reads better as a self-contained contract than as a
preamble to the code that happens to use it.

Pure move, no behaviour change: the magic, the field caps, storeLayout,
its size assertion and the two field accessors go to mp_store_layout.h,
private to the store implementation. The uid-warning throttle stays with
the reader it is policy for.
