+++ src/plugins/telemetry/prometheus_mp/mp_store.cpp
@@ -17,6 +17,7 @@
 #include "mp_store.h"
 
 #include "common/nixl_log.h"
+#include "mp_store_layout.h"
 #include "scoped_fd.h"
 
 #include <fcntl.h>
@@ -41,46 +42,6 @@ namespace nixl::telemetry::mp {
 
 namespace {
 
-    // "NIXLMPS1" as a little-endian tag; changing the layout must change either this
-    // or MP_STORE_SCHEMA_VERSION so stale-format files are rejected.
-    constexpr uint64_t MP_STORE_MAGIC = 0x3153504d4c58494eULL;
-
-    constexpr std::size_t MP_MAX_AGENT_NAME = 256;
-    constexpr std::size_t MP_MAX_HOSTNAME = 128;
-    constexpr std::size_t MP_MAX_LOCAL_RANK = 64;
-
-    // Fixed on-disk layout. Plain trivially-copyable POD operated on with __atomic
-    // builtins (not std::atomic) so it is safe to memset/reinterpret over an mmap'd
-    // region shared between processes. Field order keeps every uint64 8-byte aligned.
-    struct storeLayout {
-        uint64_t magic;
-        uint32_t schemaVersion;
-        uint32_t slotCount;
-        int64_t pid;
-        uint64_t startTime;
-        uint64_t lastUpdateNs;
-        uint64_t instance;
-        // 64-bit purely so the double array that follows stays 8-byte aligned
-        // without implicit padding.
-        uint64_t bucketCount;
-        char agentName[MP_MAX_AGENT_NAME];
-        char hostname[MP_MAX_HOSTNAME];
-        char localRank[MP_MAX_LOCAL_RANK];
-        double bucketBounds[MP_STORE_MAX_BUCKETS];
-        uint64_t counters[MP_STORE_SLOT_COUNT];
-        uint64_t gauges[MP_STORE_SLOT_COUNT];
-        uint64_t histBuckets[MP_STORE_SLOT_COUNT][MP_STORE_MAX_BUCKETS + 1];
-        uint64_t histSums[MP_STORE_SLOT_COUNT];
-    };
-
-    // schemaVersion and slotCount alone do not catch a reordered field or a
-    // changed MP_MAX_* cap, both of which move every offset after them while a
-    // peer still validates the header. Failing the build is the only way to force
-    // the version bump that makes such a file rejectable.
-    static_assert(sizeof(storeLayout) == 6808,
-                  "storeLayout is an on-disk format: bump MP_STORE_SCHEMA_VERSION when its size "
-                  "changes, then update this assertion");
-
     // A store owned by another uid is never reaped, so the same file would warn on
     // every scrape for the life of the process. Bounded so a directory full of
     // planted files cannot grow this without limit.
@@ -91,23 +52,6 @@ namespace {
         static std::set<std::filesystem::path> seen;
         const std::lock_guard<std::mutex> lock(mutex);
         return seen.size() < max_remembered && seen.insert(path).second;
-    }
-
-    void
-    copyField(char *dst, std::size_t cap, const std::string &src, const char *what) {
-        if (src.size() >= cap) {
-            NIXL_WARN << "prometheus_mp: " << what << " '" << src << "' exceeds " << (cap - 1)
-                      << " chars; truncating in telemetry store";
-        }
-        const std::size_t n = std::min(src.size(), cap - 1);
-        std::memcpy(dst, src.data(), n);
-        dst[n] = '\0';
-    }
-
-    [[nodiscard]] std::string
-    readField(const char *src, std::size_t cap) {
-        const std::size_t n = ::strnlen(src, cap);
-        return std::string(src, n);
     }
 
 } // namespace
